We hereby massacre OrthContext and its ubiquitous presence and bring instead the glorious new reign of an AS3 dependency injection framework called SwiftSuspenders. As its syntax is quite a lot like Guice, it should feel familiar. A very brief introduction can be found here:
    https://github.com/tschneidereit/SwiftSuspenders

Since _ctx is no longer the source of everything to every class, it's shrunk to a tiny fraction of its former size and while I still use it as the entry point of the client, there's no overwhelming reason to keep that either.

I deliberately mimiced the injector configuration system we use in Narya, with an OrthModule that sets things up and static run() and runClient() methods that eventually end up constructing and initializing the context class in what I hope is a subclass-friendly way.

There is one substantial drawback to this system: Flex cannot introspect on protected variables, so directly injected members must be public. While I initially balked at 'public var _port :int', by now it doesn't bother me so much. We're all very used to seeing a field that begins with _ as private to the class.

That said, I welcome all thoughts.