Eliminate aslib-config.xml and jar and swc copying.  Simpler times.
Maven ant's dependencies task lets you cache its dependency resolution in an xml file importable by
ant.  If you do that and the file exists on later runs, the dependency properties and paths are
pulled in from the xml rather than doing the actual http resolution.  This makes it feasible to run
the maven dependency builder every time to create actionscript and java compilation paths.

As long as those cache files are in place, maven doesn't check for updates to snapshots.  If you
need a new snapshot, run ant clean.