Boom! OK, here's a real proposal for how the place resolution business could work.
OrthPlace is a fully qualified type that refers to the precise location of an instantiated place (room or intervention). It's what we store in PlayerObject.location, and it's what the location/hosting method returns.

PlaceKey is what we use to tell said service which room/intervention we want hosted. As such, it needs to contain some kind of (comparable) key that the appropriate world implementation understands. For rooms, it's simply sceneId. For interventions, it's TBD, but I'm wondering if it might not simplify things to generate a 128-bit hash or something like that, so we can easily and uniquely reference a potential intervention.

For a client to move to a place, it first calls locatePlace() on its Aether connection (through WorldDirector, not yet quite implemented). The server will look among its peers if the place is already resolved; if not, it acquires a peer and resolves it locally (proper load balancing comes later).

When the client receives the OrthPlace result, it instructs its world client to connect to that OrthPlace's getPeer(). From there on it turns a little fuzzy again; we could eithe have a general SomeService.moveTo(OrthPlace) that relays the request to the implementaions, or we might hit the implementations directly, which should be easy now that the resolution is all done and the world client knows it's on the right peer.

This is still half just a sketch. Input welcome.