Hook up tells.
To get this to work, I needed to pull aether's auth name up into orth to allow the client lookup to
work. Previously tells and player actions were using AuthName.makeKey to locate client objects, but
that doesn't work as AuthName's equals and compare both use their implementation class as part of
the determination. I killed AuthName.makeKey in favor of having AuthName subclass specific makeKey
methods, which seems workable.

This does mean aether (and party) implementations all need to use the orth-level name and can't
subclass it. I don't think this is going to be a problem since auth names shouldn't be carrying any
additional functionality, but we'll see.