Unleash a plague of Locus!
This replaces orth's world system with a similar one called locus. locus is the base noun you end up
with for your location system when you've got 9 years of variations on location services in your
classpath and all the good names are taken. In addition to replacing world's reuse of place with
a new noun, locus changes a few other assumptions:

* The peer lookup of a type of loci is left up to its creator instead of being handled by locus.
  This lets rooms be shared between peers while instancing interventions.
* Loci creators are looked up via a single Locus subclass they represent rather than via an
  identifier String.
* There is no notion of a shared body object with a locus on it. Instead the subscriber-specific
  views vended by aether will publish the player's locus to interested peers. We can't assume that
  peered services will be on the locus hosting peer as the party system was doing.

Next I'll be cleaning up the client's world assumptions, but this works well enough to support
multiple loci per type.