Drop OrthSpeakProvider and rename SpeakObject to SpeakRouter
Routing to a speak object is location-specific enough that OrthSpeakProvider doesn't really buy us
anything, and not having it means SpeakObject doesn't need to do anything Java-side.

SpeakObject isn't a DObject, so giving it the Object suffix is confusing. Rename it to SpeakRouter,
and make a marker in Java with a more actionscript-y interface on the AS side.