Separate locii and nodelets
A Locus is now just an interface that can be implemented to allow a client to ask the server to host
a place. It has no notion of peering. A Nodelet is the mechanism whereby a dobject location can be
published to multiple peers for discoverability by clients. Nodelets that are also locii just
implement Locus.

This gets rid of all the weirdness in Nodelet where it could optionally implement Publishable to
indicate it could be peered. Now a Nodelet must have a key and be peerable.

This required running genstreamable, which happened to move the import out of the generated
preamble. That made for way more change than really happened.