Fix broken asdoc comments.
This is a non-comprehensive drive by change because I was tired of only seeing the description up to the first {. Apparently, anything more is not legal asdoc and/or breaks the flash builder parser.

Anyhoo, for future reference, asdoc does not support @link. In fact, there is not much that it does support. Just use straight html and @see tags: http://livedocs.adobe.com/flex/3/html/asdoc_6.html