Lay the groundwork for instancing -- work in progress.
This follows ProjectX's solution pretty closely. The Instance class consists mostly of code copied directly out of SceneRegistry, and relies on InstancingSceneRegistry and InstancedSceneManager to intercept and relay important calls and callbacks.

An alternate solution might have been possible if it'd been possible to instantiate SceneRegistry, but it's adamant about being a Singleton and rabbitholing into a Vilya refactor is outside the scope of this project.

Code review is welcome at the concept level; the details of the implemntation will see considerable tightening up in the days to come.