Away with VizPlayerName.
It was a useful abstraction for MSOY, but it misses the mark for the general case, and specifically for Who. The times we want to show a headshot, we almost certainly want to show other player attributes as well (e.g. level), so there will be a DSet<SomethingEntry> anyway.