Some authentication stuff for aether
Changed inheritance of AetherCredentials, making it much more useful for who.

On the client, AetherCredentials now automatically sets the ident.

The AetherAuthentiator ended up being just a collection of utils. I had wanted to roll in the
actual processAuthentication, but the logic turned out to be too complicated because orth knows
nothing about users and ooo-user, which handles all the banning and ident updating.