A new player-coordinating singleton.
The idea is that players have different clients that log in and out of various nodes, and it's only in aggregate that we can authoritatively say what state a player is in. We combine Aether and Locus updates and intelligently rank Whereabouts to provide client classes with a simple signal-based subscription framework.