Use a consistent way for AuthName subclasses to categorize themselves.
Previously, LocusAuthName's compareTo() was not symmetric to other AuthNames, so you could have a.compareTo(b) = 1 but b.compareTo(a) != -1, which wreaks havoc on binarySearch() and the like.

Now, getDiscriminator() returns a category that a class wants to be part of. By default it's the concrete class name, but e.g. LocusAuthName overrides it so that all its subclasses are seen as identical for lookup purposes.