Generalize SpeakRouter away from its DObject obsession.
There used to be precisely one way for Speaks to broadcast from server to clients; posting a message on DObjects. This is still the main way we do it, but to prepare for global chats, where having channel subscription equate to DObject subscription feels fragile and annoying, we want to be able to implement different types of SpeakRouters.