Parties now use the Nodelet system.
There is no immediate tangible benefit to this, but all the weird handcoded host/port/client juggling we did before was a constant source of fragility. With everything on the same footing as e.g. guilds, further improvements should be much easier.

This is a pretty major change. I had to rewrite a ton of code. Things probably broke.