Whenever a locus is materialized, update it on the caller.
I'm not sure this is how we ought to do this. For one thing, interventions bypass this entire system (though it's a separate question entirely whether or not interventions even technically qualify as using the locus framework). For another, we already have code in all the managers to update a player's whereabouts manually; this could easily be abstracted to a locus-updating thing.