A long time ago we made all LocusAuthNames equal(). That was perhaps a bit overzealous.
The purpose was to be able to use a single LocusAuthName.makeKey() to locate any type of Locus connection anywhere, through the peer nodes. But of course it had other, less desirable effects as well, to wit: connecting as RoomAuthName, and the system gladly telling you you've resumed an InterventionName connection. Not good.

So we revert this change and find another way of locating Locus bodies.