Add ignore data to OrthClientInfo.
This means every peer knows the full ignore list of every online player. This shouldn't be a whole lot of data in the average case, and since we do need to test ignoring from e.g. guild and party peers, the alternative would be lots of node actions, which would surely cost more resources.