Massive party rewrite. We now explicitly addPlayer() to parties, and then connect.
The previous approach, where the client connected to the PartyManager and sort of automatically got soaked up in the party, was really always a terrible idea, and in practice it was too fragile to endure in the brave new world of hectic multi-player teaming. We need dependable, explicit party size checking before we ask a client to connect somewhere.

The old approach didn't actually even save any resources; the NodeAction we now do to addPlayer(), the old approach did after the fact to update AetherClientObject.party.

On the Client, the main result of our change is that all party connections are slaved off of AetherClientObject.party. We don't explicitly start as connection after e.g. joinParty and createParty; instead, we wait for those server calls to update our player object, and then act on that side effect.