A function that's called on the invite when we briefly have the aether client object.
This is a bit weird. We stream the invite to the remote node, call a method on it that is allowed to modify the invite using data on that node, and then we stream it back. It seems rather excessive for its purpose. On the other hand it's sort of glorious.