A way for locus implementations to i.e. kick a player.
This in itself has no direct effect. The client doesn't react on the locus changing, nor should it. The main reason to change the locus is so that next time the player logs in, they don't try to return to wherever they were kicked from.

It's possible this should be removed by an AetherReceiver.requestMove().